+++ CHANGELOG.md
@@ -11,6 +11,16 @@ BUG FIXES:
 FEATURES:
 
 * Support StdoutLoggerf that allows control log level (resolve issue #84).
+* Support new BucketsSearchMode config to set policy for BucketDiscovery (resolve #71).
+
+REFACTOR:
+
+* Func bucketSearchLegacy: log error from bucketStatWait (except bucketStatError).
+* New bucketsDiscoveryAsync, bucketsDiscoveryWait, bucketsDiscovery methods for buckets discovery pagination.
+* Support bucketSearchBatched method for batched buckets discovery (resolve #71).
+
+TESTS:
+* Tests for BucketsSearchMode (tnt/discovery_test.go).
 
 ## v1.1.0